+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix resources exceptions routes responses, add resources exceptions tests [#176](https://github.com/datagouv/hydra/pull/176)
 - Fix CSV analysis CLI [#181](https://github.com/datagouv/hydra/pull/181)
 - Add a `PUT` `/api/resources-exceptions/{id}` route to update a resource exception [#178](https://github.com/datagouv/hydra/pull/178)
+- Add a `quiet` argument for `purge_check` and `purge_csv_table` CLIs [#184](https://github.com/datagouv/hydra/pull/184)
 
 ## 2.0.0 (2024-09-24)